+++ WordPress/Classes/Services/SiteAddressService.swift
@@ -5,13 +5,13 @@ import WordPressKit
 typealias SiteAddressServiceCompletion = (Result<[DomainSuggestion]>) -> Void
 
 protocol SiteAddressService {
-    func addresses(for query: String, segmentID: Int, completion: @escaping SiteAddressServiceCompletion)
+    func addresses(for query: String, segmentID: Int64, completion: @escaping SiteAddressServiceCompletion)
 }
 
 // MARK: - MockSiteAddressService
 
 final class MockSiteAddressService: SiteAddressService {
-    func addresses(for query: String, segmentID: Int, completion: @escaping SiteAddressServiceCompletion) {
+    func addresses(for query: String, segmentID: Int64, completion: @escaping SiteAddressServiceCompletion) {
         let result = Result.success(mockAddresses())
         completion(result)
     }
@@ -65,7 +65,7 @@ final class DomainsServiceAdapter: LocalCoreDataService, SiteAddressService {
 
     // MARK: SiteAddressService
 
-    func addresses(for query: String, segmentID: Int, completion: @escaping SiteAddressServiceCompletion) {
+    func addresses(for query: String, segmentID: Int64, completion: @escaping SiteAddressServiceCompletion) {
 
         domainsService.getDomainSuggestions(base: query,
                                             segmentID: segmentID,
